font-recognition-backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
tensorflow==2.13.0
pillow==9.5.0
numpy==1.24.3
